requirements.txt
langchain
langchain-community
langchain-huggingface
openai
faiss-cpu
sentence-transformers
numpy
pandas
python-dotenv
